+++ tests/pytests/conftest.py
@@ -221,10 +221,10 @@ def convert_kernels(kernels):
                     warnings.warn('Failed to convert transfer kernel, ' + kernel + ', skipping...')
                 else:
                     kernel = matches.group(1)
-                    binary_kernels.append(kernel)
+                    binary_kernels.append(os.path.abspath(kernel))
             except:
                 raise Exception(f"Unable to convert {path} to binary kernel")
-        updated_kernels.append(kernel)
+        updated_kernels.append(os.path.abspath(kernel))
     
     # Sort Kernels
     # Ensure that the ISIS Addendum kernel is last in case it overrides