+++ tests/pytests/test_newhorizons_drivers.py
@@ -34,7 +34,7 @@ def test_kernels():
 @pytest.mark.parametrize("image", ['lor_0034974380_0x630_sci_1'])
 def test_nhlorri_load(test_kernels, image):
     label_file = get_image_label(image, 'isis')
-    isd_str = ale.loads(label_file, props={'kernels': test_kernels[image]})
+    isd_str = ale.loads(label_file, props={'kernels': test_kernels[image], 'attach_kernels': False})
     compare_isd = image_dict[image]
     isd_obj = json.loads(isd_str)
     comparison = compare_dicts(isd_obj, compare_isd)
@@ -44,7 +44,7 @@ def test_nhlorri_load(test_kernels, image):
 @pytest.mark.parametrize("image", ['lsb_0296962438_0x53c_eng'])
 def test_nhleisa_load(test_kernels, image):
     label_file = get_image_label(image, 'isis')
-    isd_str = ale.loads(label_file, props={'kernels': test_kernels[image]})
+    isd_str = ale.loads(label_file, props={'kernels': test_kernels[image], 'attach_kernels': False})
     compare_isd = image_dict[image]
     isd_obj = json.loads(isd_str)
     comparison = compare_dicts(isd_obj, compare_isd)
@@ -54,7 +54,7 @@ def test_nhleisa_load(test_kernels, image):
 @pytest.mark.parametrize("image", ['mpf_0295610274_0x539_sci'])
 def test_nhmvic_load(test_kernels, image):
     label_file = get_image_label(image, 'isis')
-    isd_str = ale.loads(label_file, props={'kernels': test_kernels[image], 'exact_ck_times': False})
+    isd_str = ale.loads(label_file, props={'kernels': test_kernels[image], 'exact_ck_times': False, 'attach_kernels': False})
     compare_isd = image_dict[image]
 
     isd_obj = json.loads(isd_str)
@@ -64,7 +64,7 @@ def test_nhmvic_load(test_kernels, image):
 @pytest.mark.parametrize("image", ['mc3_0034948318_0x536_sci_1'])
 def test_nhmvictdi_load(test_kernels, image):
     label_file = get_image_label(image, 'isis')
-    isd_str = ale.loads(label_file, props={'kernels': test_kernels[image]})
+    isd_str = ale.loads(label_file, props={'kernels': test_kernels[image], 'attach_kernels': False})
     compare_isd = image_dict[image]
     isd_obj = json.loads(isd_str)
     assert compare_dicts(isd_obj, compare_isd) == []